changing pin type to uint32_t and fixing pwm_is_enabled on the API interface

+++ hw/drivers/pwm/pwm_nrf52/src/pwm_nrf52.c
@@ -439,7 +439,7 @@ nrf52_pwm_configure_channel(struct pwm_dev *dev,
         return (EINVAL);
     }
 
-    config->output_pins[cnum] = cfg->pin;
+    config->output_pins[cnum] = (uint8_t) cfg->pin;
     config->output_pins[cnum] |= (cfg->inverted) ?
         NRFX_PWM_PIN_INVERTED : config->output_pins[cnum];
